add stakeholder's Layout  and navigations
