+++ src/components/Layout/LayoutUser.jsx
@@ -1,0 +1,12 @@
+import NavBarUser from "../NavigationBar/User/NavBarUser"
+import Footer from "../Footer/Footer"
+
+const LayoutUser = ({ children }) => (
+    <>
+        <NavBarUser />
+        {children}
+        <Footer />
+    </>
+)
+
+export default LayoutUser
