--- src/components/NavigationBar/NavigationTitle.js
@@ -1,32 +1,0 @@
-export const homeTitle = [
-    {
-        key: 1,
-        title: "MENU1",
-        path: ""
-    },
-    {
-        key: 2,
-        title: "MENU2",
-        path: ""
-    },
-    {
-        key: 3,
-        title: "MENU3",
-        path: ""
-    },
-    {
-        key: 4,
-        title: "MENU4",
-        path: ""
-    },
-    {
-        key: 5,
-        title: "MENU5",
-        path: ""
-    },
-    {
-        key:6,
-        title: "MENU6",
-        path: ""
-    }
-]
